+++ bin/server.js
@@ -9,7 +9,7 @@ const net  = require('net');
 const zlib = require('zlib');
 const { execFile } = require('child_process');
 
-const Game   = require('../lib/game');
+const Game   = require('@kobalab/majiang-ai/preset-game');
 const Player = require('../lib/player');
 
 function get_shan(filename) {
@@ -24,12 +24,16 @@ function get_rule(filename = '{}') {
     return Majiang.rule(JSON.parse(fs.readFileSync(filename)));
 }
 
-function make_player(bot, callback) {
+function make_player(bot, callback, noexec) {
     const server = net.createServer((sock)=>{
         server.close();
         callback(sock);
     }).listen(()=>{
         const port = server.address().port;
+        if (noexec) {
+            console.error(bot,`mjsonp://127.0.0.1:${port}/default`);
+            return;
+        }
         execFile(bot, [`mjsonp://127.0.0.1:${port}/default`])
             .on('error', (err)=>{ throw err });
     });
@@ -52,15 +56,16 @@ const argv = require('yargs')
     .option('skip',     { alias: 's', description: '指定した数の牌山をスキップ' } )
     .option('rule',     { alias: 'r', description: 'ルール' })
     .option('verbose',  { alias: 'v', boolean: true })
+    .option('noexec',   { alias: 'X', boolean: true })
     .demandCommand(2)
     .argv;
 
-const script = get_shan(argv.input) || [];
-for (let i = 0; i < (argv.skip || 0); i++) script.shift()
+const shan = get_shan(argv.input) || [];
+for (let i = 0; i < (argv.skip || 0); i++) shan.shift()
 
 const rule = get_rule(argv.rule);
 
-let times = argv.times || script && script.length || 1;
+let times = argv.times || shan.length || 1;
 
 const bots = [ argv._[1], argv._[0], argv._[0], argv._[0] ];
 let players = [];
@@ -71,20 +76,19 @@ console.log(`[${times}]`, new Date().toLocaleTimeString());
 
 function start_game() {
     players = [];
-    let s = script.shift();
     for (let id = 0; id < 4; id++) {
         make_player(bots[id], (sock)=>{
             players[id] = new Player(sock);
             if (players.filter(s => s).length == 4) {
                 players[0].debug = argv.verbose;
-                const game = s ? new Game(players, end_game, rule).script(s)
-                               : new Majiang.Game(players, end_game, rule);
+                const game = new Game(players, end_game, rule)
+                                                    .preset(shan.shift());
                 game.model.player = player_name(game.model.player, bots);
                 game.model.title += ` #${logs.length + (argv.skip || 0)}`;
                 game.speed = 0;
                 game.kaiju();
             }
-        });
+        }, argv.noexec && id == 0);
     }
 }