+++ ChangeLog.md
@@ -1,3 +1,9 @@
+## v0.2.0 / 2026-09-25
+
+  - オプション --noexec を追加
+  - プロトコル変換に @kobalab/mjai-bot/convert を使うよう修正
+  - デュプリケート対局に @kobalab/majiang-ai/preset-game を使うよう修正
+
 ### v0.1.1 / 2026-09-23
 
   - npm に登録